+++ src/lib/components/select-code/CodeSelectSection.tsx
@@ -5,7 +5,7 @@ import type { Control, FieldPath, FieldValues } from "react-hook-form";
 import { AmpEvent, track } from "lib/amplitude";
 import { ControllerInput } from "lib/components/forms";
 import type { FormStatus } from "lib/components/forms";
-import type { CodeInfoResponseLcd } from "lib/services/types";
+import type { Code } from "lib/services/types";
 import type { Option } from "lib/types";
 
 import { CodeSelect } from "./CodeSelect";
@@ -16,7 +16,7 @@ interface CodeSelectSectionProps<T extends FieldValues> {
   control: Control<T>;
   error: Option<string>;
   onCodeSelect: (codeId: string) => void;
-  setCodeHash: (data: CodeInfoResponseLcd) => void;
+  setCodeHash: (data: Code) => void;
   status: FormStatus;
 }