+++ src/lib/pages/validators/components/ValidatorsBodyFull.tsx
@@ -75,6 +75,7 @@ export const ValidatorsBodyFull = ({
         isDesc={isDesc}
         setIsDesc={setIsDesc}
         scrollComponentId={scrollComponentId}
+        isSearch={!!search}
       />
       {data && data.total > 10 && (
         <Pagination